Adding StopsTable
