+++ md/templates/md/agency_detail.html
@@ -80,6 +80,7 @@
 
 new MD.StopRatioDonut({handler: stop_handler, selector: "#stop_race_pie"});
 new MD.StopRatioTimeSeries({handler: stop_handler, selector: "#stop_race_line"});
+new MD.StopsTable({handler: stop_handler, selector: "#stop_race_data"});
 
 {% comment %}</script>{% endcomment %}
 {% endblock graph-code %}
